More auto focus support + UI fixes
Added support for devices that do not support
“FOCUS_MODE_CONTINUOUS_VIDEO” by manually telling the camera to auto
focus every 5 seconds (devices that do support continuous focus are
unaffected).

Fixed a problem with the marker details activity where the title and
button would appear on top of each other (on a Nexus S).

Made the mode selector clickable (e.g. click on “Threshold” and it goes
to threshold mode) as a user did not know you could swipe to change
mode (now you can do both).