Fix for crash on sorting experiences
The previous experience Comparator violated transitivity by returning 0 if any parameter was null.
e.g. if A=Apple, B=null, C=Orange, the previous Comparator would say A==B, B==C, but A<B.
If the list of experiences contains null names and is ordered in a way that the runtime notices this problem the app crashes.