Experience/screen change bug fix
Fixed a bug where the experience would change when returning to the
scanning screen when using the “up/back” button in the top left corner
of the screen. Using the back button at the bottom of the screen
produced the correct behaviour.

Fix: The current experience is now saved to preferences when changed
to. An experience is loaded from preferences in onResume and when the
experience list updates.