Image processing update
Includes colour filters, white balancing and hue/lightness/saturation
editing.
ImageBuffers has been changed so that it manages the colour conversions
and keeps track of where the result of the most recent processing is.
The number of buffers/Mats is the same they’ve just been renamed (e.g.
temp is now bgrBuffer).