Fix for cacheing 2
Changed an adapter to wait for all experiences to load if loading is
triggered when it’s empty (e.g. the first time the Activity is
displayed) as otherwise the experiences loading out of order can cause
the scroll position to be moved with the first item to be loaded
starting the user in the middle of the list.